+++ common/types/BorderRadius.ts
@@ -1,0 +1,3 @@
+import { borderRadius } from '../tokens/borderRadius';
+
+export type BorderRadius = keyof typeof borderRadius;