+++ web/buttons/interactableStyles.ts
@@ -1,8 +1,6 @@
 // import { opacityDisabled } from '@cbhq/cds-common/tokens/interactableOpacity';
 import { css } from 'linaria';
 
-import { palette } from '../tokens';
-
 // TODO: Fix when we can do deep imports in common
 const opacityDisabled = 0.38;
 
@@ -12,29 +10,25 @@ export const interactable = css`
   cursor: pointer;
   user-select: none;
   text-decoration: none;
-  z-index: 0;
+  margin: 0;
   transition: transform 100ms;
   /* Prevents layout shift - https://web.dev/cls/#animations-and-transitions */
   transform: scale(1);
-
-  &:before,
-  &:after {
-    content: '';
-    position: absolute;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-  }
-
-  &:before {
-    z-index: -1;
-    opacity: var(--interactable-opacity);
-  }
-
-  &:after {
-    background-color: var(--interactable-underlay);
-    z-index: -2;
+  opacity: var(--interactable-opacity);
+  background-image: linear-gradient(
+    to right,
+    var(--interactable-background),
+    var(--interactable-background)
+  );
+  background-color: var(--interactable-underlay);
+  /* stylelint-disable plugin/no-low-performance-animation-properties */
+  transition: color 150ms ease-in-out, background-color 150ms ease-in-out,
+    background-image 150ms ease-in-out, transform 100ms;
+  /* Removes weird bonus padding in Firefox */
+  &::-moz-focus-inner {
+    border: 0;
+    padding: 0;
+    margin: 0;
   }
 `;
 
@@ -42,16 +36,7 @@ export const scaledDownState = css`
   transform: scale(0.98);
 `;
 
-export const pressedState = css`
-  --interactable-underlay: ${palette.foreground};
-`;
-
-export const hoveredState = css`
-  --interactable-underlay: ${palette.background};
-`;
-
 export const disabledState = css`
-  --interactable-underlay: unset;
   opacity: ${opacityDisabled};
   cursor: default;
   pointer-events: none;