+++ web/navigation/Navbar.tsx
@@ -1,12 +1,11 @@
 import React, { memo } from 'react';
 
-import { join } from '@cbhq/cds-common';
+import { join, gutter } from '@cbhq/cds-common';
 import { emptyArray } from '@cbhq/cds-utils';
 
 import { Box } from '../layout/Box';
 import { HStack } from '../layout/HStack';
 import { Spacer } from '../layout/Spacer';
-import { gutter } from '../tokens';
 import { TextTitle3, TextLabel2 } from '../typography';
 
 const toArray = React.Children.toArray;