+++ web/styles/borderRadius.ts
@@ -1,0 +1,21 @@
+/**
+ * DO NOT MODIFY
+ * Generated from scripts/codegen/main.ts
+ */
+import { css } from 'linaria';
+
+export const compact = css`
+  border-radius: 4px;
+`;
+
+export const standard = css`
+  border-radius: 8px;
+`;
+
+export const pill = css`
+  border-radius: 18px;
+`;
+
+export const round = css`
+  border-radius: 50%;
+`;